requirements.txt
behave
requests
pyhamcrest
allure-behave